requirements.txt
asyncio
cbor2
pynacl
numpy
pyyaml
pandas
matplotlib

